Use double instead of float for coordinates

+++ src/NGrib/Coordinate.cs
@@ -4,13 +4,13 @@ namespace NGrib
 {
 	public readonly struct Coordinate
 	{
-		public float Latitude { get; }
-		public float Longitude { get; }
+		public double Latitude { get; }
+		public double Longitude { get; }
 
-		public Coordinate(float latitude, float longitude)
+		public Coordinate(double latitude, double longitude)
 		{
-			Latitude = -90f <= latitude && latitude <= 90 ? latitude : SimplifyLatitude(latitude);
-			Longitude = longitude >= 0 && longitude < 360 ? longitude : SimplifyLongitude(longitude);
+			Latitude = -90d <= latitude && latitude <= 90d ? latitude : SimplifyLatitude(latitude);
+			Longitude = longitude >= 0d && longitude < 360d ? longitude : SimplifyLongitude(longitude);
 		}
 
 		public bool Equals(Coordinate other)
@@ -31,14 +31,14 @@ namespace NGrib
 			}
 		}
 
-		public Coordinate Add(float latitudeIncrement, float longitudeIncrement)
+		public Coordinate Add(double latitudeIncrement, double longitudeIncrement)
 		{
 			return new Coordinate(Latitude + latitudeIncrement, Longitude + longitudeIncrement);
 		}
 
-		private static float SimplifyLatitude(float latitude) => Math.Abs(latitude % 180) < 0.0000001 ? 0 : (float) Math.Asin(Math.Sin(latitude.ToRadians())).ToDegrees();
+		private static double SimplifyLatitude(double latitude) => Math.Abs(latitude % 180) < 0.0000001 ? 0 : Math.Asin(Math.Sin(latitude.ToRadians())).ToDegrees();
 
-		private static float SimplifyLongitude(float longitude) => longitude >= 0 ? longitude % 360 : 360 + longitude % 360;
+		private static double SimplifyLongitude(double longitude) => longitude >= 0 ? longitude % 360 : 360 + longitude % 360;
 
 		public override string ToString()
 		{
